# Trie HexPrefix optimizations using Raw Path