This is an experimental PR that changes how `HexPrefix` works with paths. Instead of decoding them to nibbles when reading `RPL` and encoding them back when encoding, it works with raw encoded data. Highly experimental stuff.

Initial benchmark, that does full cycle of RLP -> decode -> extend path by concat -> encode -> RLP looks really promising allocation wise and good performance wise.

|  Method |     Mean |    Error |   StdDev |   Median | Ratio | RatioSD |  Gen 0 | Allocated |
|-------- |---------:|---------:|---------:|---------:|------:|--------:|-------:|----------:|
| RawPath | 36.16 ns | 0.550 ns | 0.430 ns | 36.27 ns |  0.89 |    0.02 | 0.0051 |      64 B |
| Current | 41.17 ns | 0.793 ns | 1.410 ns | 40.50 ns |  1.00 |    0.00 | 0.0134 |     168 B |

## Changes:

- [ ] TBD ...

## Types of changes

What types of changes does your code introduce?
_Put an `x` in the boxes that apply_

- [ ] Bugfix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] Documentation Update
- [ ] Code style update (formatting, renaming)
- [x] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Other (please describe): 

## Testing
**Requires testing**

- [x] Yes
- [ ] No

**In case you checked yes, did you write tests??**

- [x] Yes
- [ ] No